+++ proxy_protocols.py
@@ -1,5 +1,4 @@
 import base64
-import ipaddress
 import json
 from urllib.parse import parse_qs, unquote, urlparse
 
@@ -83,20 +82,8 @@ def reality_fingerprint(value):
     return fingerprint
 
 
-def _is_ip_address(value):
-    try:
-        ipaddress.ip_address(str(value or ''))
-    except ValueError:
-        return False
-    return True
-
-
 def vless_outbound_address(data):
-    address = data.get('address') or data.get('host', '')
-    server_name = data.get('sni') or data.get('host', '')
-    if data.get('security') == 'reality' and _is_ip_address(address) and server_name and not _is_ip_address(server_name):
-        return server_name
-    return address
+    return data.get('address') or data.get('host', '')
 
 
 def parse_trojan_key(key):
@@ -181,7 +168,7 @@ def proxy_outbound_from_key(proto, key_value, tag, email='t@t.tt'):
         if tls_mode in ['tls', 'xtls']:
             stream_settings['security'] = tls_mode
             stream_settings[f'{tls_mode}Settings'] = {
-                'serverName': data.get('add', ''),
+                'serverName': data.get('sni') or data.get('host') or data.get('add', ''),
             }
         else:
             stream_settings['security'] = 'none'